+++ src/test/kotlin/com/quizit/quiz/util/TestUtil.kt
@@ -1,32 +1,7 @@
 package com.quizit.quiz.util
 
-import com.quizit.quiz.domain.enum.Role
-import com.quizit.quiz.fixture.createJwtAuthentication
-import org.springframework.restdocs.payload.FieldDescriptor
-import org.springframework.restdocs.payload.PayloadDocumentation.fieldWithPath
-import org.springframework.restdocs.request.ParameterDescriptor
-import org.springframework.restdocs.request.RequestDocumentation.parameterWithName
-import org.springframework.security.core.authority.SimpleGrantedAuthority
-import org.springframework.security.core.context.SecurityContextHolder
+import org.reactivestreams.Publisher
+import reactor.test.StepVerifier
+import reactor.test.StepVerifier.FirstStep
 
-infix fun String.desc(description: String): FieldDescriptor =
-    fieldWithPath(this)
-        .description(description)
-
-infix fun String.paramDesc(description: String): ParameterDescriptor =
-    parameterWithName(this)
-        .description(description)
-
-fun withMockUser() {
-    SecurityContextHolder.getContext().authentication = createJwtAuthentication()
-}
-
-fun withMockAdmin() {
-    SecurityContextHolder.getContext().authentication =
-        createJwtAuthentication(authorities = listOf(SimpleGrantedAuthority(Role.ADMIN.name)))
-}
-
-val errorResponseFields = listOf(
-    "code" desc "상태 코드",
-    "message" desc "에러 메세지"
-)
+fun <T> Publisher<T>.getResult(): FirstStep<T> = StepVerifier.create(this)
